Mostly Working Windows Support
The windows support is mostly working and the command line interface is
functioning. I still need to test all the modules on Linux again to make
sure architecture changes haven't broken anything there. Also, more work
is needed on the Windows C2 to solidify the API, but it is working at
it's core.