Fixed a possible exception leak
Previously, we were registering a new 'session' with the 'manager' before completing the setup: '__init__'. Any exception raised after we registered the 'session' didn't remove the registered session. This will allow the reuse of same 'session_id' because we have already registered the 'session' with this 'session_id', which may possibly be an invalid session. And 'session_id' variable in 'manager' will not be updated if an exception is raised. Thus, I have shifted down the session registering part so that it occurs after the completion of session creation.