Added 'refresh_uid()' call, more checks for user/group ids
If the attacker makes use of a SUID or SGID binary and their 'id' updates, then we need to refresh the 'id' in pwncat's internals as well, which is done during the session initialization.
Add '-p' flag for 'sh' command so that it does not drop any 'euid' or 'egid' updates.
Fixed some minor typos, and added a 'docstring' for 'writable' method.